+++ CHANGELOG.md
@@ -12,8 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Class renamed:
   - `GenericCardSelection` -> `GenericCardSelectionExtension`
-- Methods renamed and signatures refactored:
+- Methods renamed or signatures refactored:
   - `GenericCardSelection createCardSelection ()` -> `GenericCardSelectionExtension createGenericCardSelectionExtension()`
+  - `CardResourceProfileExtension createCardResourceProfileExtension(GenericCardSelectionExtension 
+  genericCardSelectionExtension)`  -> `CardResourceProfileExtension createCardResourceProfileExtension
+  (CardSelector<IsoCardSelector> cardSeletor, GenericCardSelectionExtension genericCardSelectionExtension)`
   - `processApdusToByteArrays ()` -> `processApdusToByteArrays (ChannelControl channelControl)`
     and `processApdusToHexStrings ()` -> `processApdusToHexStrings (ChannelControl channelControl)
     The enum `ChannelControl` has been created for this purpose.